+++ lib/http/intercepted_client.dart
@@ -290,7 +290,7 @@ class InterceptedClient extends BaseClient {
     } on Exception catch (error) {
       if (retryPolicy != null &&
           retryPolicy!.maxRetryAttempts > _retryCount &&
-          retryPolicy!.shouldAttemptRetryOnException(error, request)) {
+          await retryPolicy!.shouldAttemptRetryOnException(error, request)) {
         _retryCount += 1;
         return _attemptRequest(request);
       } else {